file upload handling. Files now uploaded into directory and entry is added to database. Same deleting files

+++ application/views/admin/upload.php
@@ -11,7 +11,10 @@
 <div class="admin-container claro">
     <h2>Upload files</h2>
 
-    <?php echo $error; ?>
+    <?php
+        $class = isset($error) ? 'error' : 'success';
+        echo "<span class='$class'>" . $message . "</span>";
+    ?>
 
     <?php echo form_open_multipart('/admin/do_upload');?>
         <label for="description">Description of File</label>
@@ -24,11 +27,20 @@
     <h2>All uploaded files</h2>
     <table>
         <?php
-            foreach ($files as $file) {
-                echo "<tr>";
-                echo "<td>" . $file->filename . " </td>";
-                echo "<td>" . $file->description . " </td>";
-                echo "<td><a href='/admin/delete_file?id=" . $file->id . "'>delete</a></td>";
+            $files = $this->file->get_all();
+            if ($files) {
+                foreach ($files as $file) {
+                    echo "<tr>";
+                    echo "<td>" . $file->filename . " </td>";
+                    echo "<td>" . $file->description . " </td>";
+                    echo "<td><a href='/admin/delete_file?id=" . $file->id . "'>delete</a></td>";
+                    echo "</tr>";
+                }
+            } else {
+                echo "<tr class='no-files'>";
+                echo "<td><i>No files uploaded</i></td>";
+                echo "<td></td>";
+                echo "<td></td>";
                 echo "</tr>";
             }
         ?>
